# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A search engine API that aggregates content from multiple providers, ranks them by relevance score, and provides search, filtering, sorting, and pagination capabilities.",
        "title": "Search Engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "rezanicgil@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/content/{id}": {
            "get": {
                "description": "Get detailed information about a specific content item by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get content by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Content ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Content"
                        }
                    },
                    "400": {
                        "description": "Invalid content ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Content not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Content": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "integer"
                },
                "created_at": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "provider": {
                    "description": "Provider information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Provider"
                        }
                    ]
                },
                "provider_id": {
                    "type": "integer"
                },
                "published_at": {
                    "description": "Common fields",
                    "type": "string"
                },
                "reactions": {
                    "type": "integer"
                },
                "reading_time": {
                    "description": "Article-specific metrics (NULL for videos)\nThese fields are populated when Type is \"article\"",
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "tags": {
                    "description": "Related data (loaded separately)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.ContentType"
                },
                "updated_at": {
                    "type": "string"
                },
                "views": {
                    "description": "Video-specific metrics (NULL for articles)\nThese fields are populated when Type is \"video\"",
                    "type": "integer"
                }
            }
        },
        "model.ContentType": {
            "type": "string",
            "enum": [
                "video",
                "article"
            ],
            "x-enum-varnames": [
                "ContentTypeVideo",
                "ContentTypeArticle"
            ]
        },
        "model.Provider": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "format": {
                    "$ref": "#/definitions/model.ProviderFormat"
                },
                "id": {
                    "type": "integer"
                },
                "last_fetched_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rate_limit_per_minute": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.ProviderFormat": {
            "type": "string",
            "enum": [
                "json",
                "xml"
            ],
            "x-enum-varnames": [
                "ProviderFormatJSON",
                "ProviderFormatXML"
            ]
        },
...
    }
}